Remove the Internal Package
- Remove the Internal Package to make them discoverable by third party
  plugins.
- Merge de.marw.cdt.cmake.core.internal with de.marw.cdt.cmake.core.
- Move all classes from de.marw.cdt.cmake.core.internal.settings to
  de.marw.cdt.cmake.core.settings.
- Move all classes from de.marw.cdt.cmake.core.internal.storage to
  de.marw.cdt.cmake.core.storage.
- Update all references to the internal package to the new package
  names.
- Update inline comments not automatically refactored by Eclipse.
- Move internal tests into the new packages.
