+++ de.marw.cdt.cmake.core/src/main/java/de/marw/cdt/cmake/core/CmakeGenerator.java
@@ -8,7 +8,7 @@
  * Contributors:
  *      Martin Weber - Initial implementation
  *******************************************************************************/
-package de.marw.cdt.cmake.core.internal;
+package de.marw.cdt.cmake.core;
 
 /**
  * Represents a cmake build-script generator including information about the
