+++ de.marw.cdt.cmake.core/src/main/java/de/marw/cdt/cmake/core/settings/CMakePreferences.java
@@ -8,7 +8,7 @@
  * Contributors:
  *      Martin Weber - Initial implementation
  *******************************************************************************/
-package de.marw.cdt.cmake.core.internal.settings;
+package de.marw.cdt.cmake.core.settings;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,9 +16,9 @@ import java.util.List;
 import org.eclipse.cdt.core.settings.model.ICStorageElement;
 
 import de.marw.cdt.cmake.core.CdtPlugin;
-import de.marw.cdt.cmake.core.internal.storage.CmakeDefineSerializer;
-import de.marw.cdt.cmake.core.internal.storage.CmakeUndefineSerializer;
-import de.marw.cdt.cmake.core.internal.storage.Util;
+import de.marw.cdt.cmake.core.storage.CmakeDefineSerializer;
+import de.marw.cdt.cmake.core.storage.CmakeUndefineSerializer;
+import de.marw.cdt.cmake.core.storage.Util;
 
 /**
  * Holds per-configuration project settings.
