+++ de.marw.cdt.cmake.core/src/main/java/de/marw/cdt/cmake/core/settings/ConfigurationManager.java
@@ -8,7 +8,7 @@
  * Contributors:
  *      Martin Weber - Initial implementation
  *******************************************************************************/
-package de.marw.cdt.cmake.core.internal.settings;
+package de.marw.cdt.cmake.core.settings;
 
 import java.util.WeakHashMap;
