+++ de.marw.cdt.cmake.core/src/main/java/de/marw/cdt/cmake/core/storage/StorageSerializer.java
@@ -8,7 +8,7 @@
  * Contributors:
  *      Martin Weber - Initial implementation
  *******************************************************************************/
-package de.marw.cdt.cmake.core.internal.storage;
+package de.marw.cdt.cmake.core.storage;
 
 import org.eclipse.cdt.core.settings.model.ICStorageElement;
