+++ de.marw.cdt.cmake.core/src/main/java/de/marw/cdt/cmake/core/ui/LinuxPropertyTab.java
@@ -12,9 +12,9 @@ package de.marw.cdt.cmake.core.ui;
 
 import java.util.EnumSet;
 
-import de.marw.cdt.cmake.core.internal.CmakeGenerator;
-import de.marw.cdt.cmake.core.internal.settings.CMakePreferences;
-import de.marw.cdt.cmake.core.internal.settings.LinuxPreferences;
+import de.marw.cdt.cmake.core.CmakeGenerator;
+import de.marw.cdt.cmake.core.settings.CMakePreferences;
+import de.marw.cdt.cmake.core.settings.LinuxPreferences;
 
 /**
  * UI to control host Linux specific project properties and preferences for
@@ -28,7 +28,7 @@ public class LinuxPropertyTab extends AbstractOsPropertyTab<LinuxPreferences> {
       .of(CmakeGenerator.UnixMakefiles,CmakeGenerator.Ninja);
 
   /*-
-   * @see de.marw.cdt.cmake.core.ui.AbstractOsPropertyTab#getOsPreferences(de.marw.cdt.cmake.core.internal.CMakePreferences)
+   * @see de.marw.cdt.cmake.core.ui.AbstractOsPropertyTab#getOsPreferences(de.marw.cdt.cmake.core.settings.CMakePreferences)
    */
   @Override
   protected LinuxPreferences getOsPreferences(CMakePreferences prefs) {
