+++ de.marw.cdt.cmake.core/src/main/java/de/marw/cdt/cmake/core/ui/UnDefinesViewer.java
@@ -43,7 +43,7 @@ import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.Table;
 import org.eclipse.swt.widgets.TableColumn;
 
-import de.marw.cdt.cmake.core.internal.settings.CmakeUnDefine;
+import de.marw.cdt.cmake.core.settings.CmakeUnDefine;
 
 /**
  * Displays a table for the Cmake undefines. The created table will be displayed
